src/microservices/requirements.txt
flask==2.3.2
flask-cors==4.0.0
pandas==2.0.2
openpyxl==3.1.2
redis==4.5.5
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Pillow==9.5.0
requests==2.31.0

